chore: some controller

+++ src/controllers/student_controller.rs
@@ -172,22 +172,14 @@ pub async fn update_student(body: web::Json<Value>, arg: web::Path<i32>) -> impl
 
   let dob = NaiveDate::parse_from_str(date_of_birth.as_str(), "%Y-%m-%d").unwrap();
   let data = sqlx::query_as!(StudentStruct,
-    "update students set first_name = $1, last_name = $2, email = $3, phone = $4, date_of_birth = $5, region = $6, register_number = $7, toefl_score = $8, ielts_score = $9
-      where id = $10 returning *",
+    "update students set first_name = $1, last_name = $2, email = $3, phone = $4, date_of_birth = $5,
+      region = $6, register_number = $7, toefl_score = $8, ielts_score = $9 where id = $10 returning *",
     first_name, last_name, email, phone, dob, region, register_number, toefl_score, ielts_score, id)
     .fetch_all(&pool)
     .await
     .unwrap();
 
   let result = convert_vec_to_values(data);
-
-  if check_if_empty(result.to_owned()) {
-    return response_json(
-      "failed".to_string(),
-      "Student not found".to_string(),
-      vec![]
-    )
-  }
 
   response_json(
     "success".to_string(),
@@ -199,26 +191,20 @@ pub async fn update_student(body: web::Json<Value>, arg: web::Path<i32>) -> impl
 #[doc = "Delete student by id"]
 pub async fn delete_student(arg: web::Path<i32>) -> impl Responder {
   let id = arg.to_owned();
-
   let pool = connect_postgres().await;
+  let student_exists = sqlx::query_scalar::<_, bool>("select exists(select 1 from students where id = $1) as student_exists")
+    .bind(id.clone())
+    .fetch_one(&pool)
+    .await
+    .unwrap_or(false);
 
-  match sqlx::query!("select id from students where id = $1 limit 1", id.clone())
-    .fetch_optional(&pool)
-    .await {
-      Ok(Some(_)) => (),
-      Ok(None) => {
-        return response_json(
-          "failed".to_string(),
-          "Student not found".to_string(),
-          vec![]
-        )
-      }
-      Err(_) => return response_json(
-        "error".to_string(),
-        "Something went wrong".to_string(),
-        vec![]
-      )
-    };
+  if !student_exists {
+    return response_json(
+      "failed".to_string(),
+      "University not found".to_string(),
+      vec![]
+    )
+  }
 
   let data = sqlx::query_as!(StudentStruct, "delete from students where id = $1 returning *", id)
     .fetch_all(&pool)
@@ -226,14 +212,6 @@ pub async fn delete_student(arg: web::Path<i32>) -> impl Responder {
     .unwrap();
 
   let result = convert_vec_to_values(data);
-
-  if check_if_empty(result.to_owned()) {
-    return response_json(
-      "failed".to_string(),
-      "Student not found".to_string(),
-      vec![]
-    )
-  }
 
   response_json(
     "success".to_string(),
